+++ src/main/java/com/sparta/newsfeedapp/dto/comment/CommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeedapp.dto.commentResponseDto;
+package com.sparta.newsfeedapp.dto.comment;
 
 import com.sparta.newsfeedapp.entity.Comment;
 import jakarta.persistence.Column;
