+++ src/main/java/com/sparta/newsfeedapp/dto/user/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeedapp.dto.userResponseDto;
+package com.sparta.newsfeedapp.dto.user;
 
 import lombok.Getter;
