+++ src/main/java/com/sparta/newsfeedapp/dto/user/deleteRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeedapp.dto.userRequestDto;
+package com.sparta.newsfeedapp.dto.user;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
